dca: Integer core decoder.
Provide a bit-exact reconstruction.

The dca core decoder converts integer coefficients read from the
bitstream to floats just after reading them (along with dequantization).
All the other steps of the audio reconstruction are done with floats
which makes the output for the DTS lossless extension (XLL)
actually lossy.
This patch trasforms the dca core to work with integer coefficients
till the point QMF filters are called. The integer coefficients are
transformed to floats at this point if needed, the new bitexact QMF
filters are used for the XLL streams or for the case the new
-force_fixed option is set. This option forces fixed-point
reconstruction for any kind of input.